feat: :bug: Achievement all display bug solved
separate achievement list
