+++ frontend/src/pages/ProfilePage/index.tsx
@@ -7,7 +7,7 @@ import { useGetUser } from 'hooks';
 import { ProfileCard } from './ProfileCard';
 import { EditProfile } from './EditProfile';
 import { ManageFriends } from './ManageFriends';
-import { Achievement } from './Achievement';
+import { AchievementList } from './AchievementList';
 
 export const ProfilePage = () => {
   const [isModalShown, setModalShown] = useState<Boolean>(false);
@@ -35,7 +35,7 @@ export const ProfilePage = () => {
           </button>
         )}
         {!(!id || profile.id === myProfile.id) && <ManageFriends user={profile} />}
-        <Achievement userId={profile.id} />
+        <AchievementList userId={profile.id} />
       </main>
       {isModalShown && (
         <Modal onClickClose={handleCloseModal}>
